Update apiProperties.json
I think crm.objects.companies.read - crm.schemas.companies.read are required for this App to install without the error 
"Error on connection - Couldn’t complete the connection Insufficient scopes were provided." As seen when trying to add or use this Connector in a Flow.

I would love to test this, but not 100% sure how, Adding these two scopes to my Private App in Hubspot, and using the apiDefinition.swagger.json in a Custom Connector appears to work. I will continue to try and add Search functionality to the customer connector and share it if I am successful.